+++ MLEM.Data/Json/JsonTypeSafeGenericDataHolder.cs
@@ -1,5 +1,4 @@
 ﻿using System.Collections.Generic;
-using System.Runtime.Serialization;
 using MLEM.Misc;
 using Newtonsoft.Json;
 
@@ -8,7 +7,6 @@ namespace MLEM.Data.Json {
     /// An <see cref="IGenericDataHolder"/> represents an object that can hold generic key-value based data.
     /// This class uses <see cref="JsonTypeSafeWrapper"/> for each object stored to ensure that objects with a custom <see cref="JsonConverter"/> get deserialized as an instance of their original type if <see cref="JsonSerializer.TypeNameHandling"/> is not set to <see cref="TypeNameHandling.None"/>.
     /// </summary>
-    [DataContract]
     #if NET7_0_OR_GREATER
     [System.Diagnostics.CodeAnalysis.RequiresDynamicCode("The native code for instantiation of JsonTypeSafeWrapper instances might not be available at runtime.")]
     #endif
@@ -16,7 +14,7 @@ namespace MLEM.Data.Json {
 
         private static readonly string[] EmptyStrings = new string[0];
 
-        [DataMember(EmitDefaultValue = false)]
+        [JsonProperty]
         private Dictionary<string, JsonTypeSafeWrapper> data;
 
         /// <inheritdoc />